requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.14
django-dirtyfields==1.9.3
djangorestframework==3.15.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
factory-boy==3.3.0
Faker==26.0.0
flake8==7.1.0
idna==3.7
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pycodestyle==2.12.0
pyflakes==3.2.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
